requirements.txt
fastapi
httpx
uvicorn
aiosqlite
python-dotenv
apscheduler
filelock
tzdata